✨feat: 게시글 댓글 구현 중

+++ snail-member/src/main/kotlin/com/bockerl/snailmember/file/command/domain/service/CommandFileServiceImpl.kt
@@ -15,6 +15,7 @@ import com.bockerl.snailmember.file.command.domain.aggregate.entity.File
 import com.bockerl.snailmember.file.command.domain.aggregate.vo.CommandFileRequestVO
 import com.bockerl.snailmember.file.command.domain.aggregate.vo.CommandFileWithGatheringRequestVO
 import com.bockerl.snailmember.file.command.domain.repository.CommandFileRepository
+import io.github.oshai.kotlinlogging.KotlinLogging
 import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
 import org.springframework.web.multipart.MultipartFile
@@ -27,8 +28,11 @@ class CommandFileServiceImpl(
     private val commandFileRepository: CommandFileRepository,
     private val gatheringFileService: CommandGatheringFileService,
 ) : CommandFileService {
+
+    private val logger = KotlinLogging.logger {}
+
     @Transactional
-    override fun uploadProfileImage(
+    override fun uploadSingleFile(
         file: MultipartFile,
         commandFileRequestVO: CommandFileRequestVO,
     ) {
@@ -283,12 +287,14 @@ class CommandFileServiceImpl(
                 commandFileRequestVO.fileTargetId,
             )
 
-        for (file in files) {
-            val blobClient = blobContainerClient.getBlobClient(file.fileName)
-            blobClient.delete()
-        }
+        if (files.isNotEmpty()) {
+            for (file in files) {
+                val blobClient = blobContainerClient.getBlobClient(file.fileName)
+                blobClient.delete()
+            }
 
-        commandFileRepository.deleteAll(files)
+            commandFileRepository.deleteAll(files)
+        }
     }
 
     /* 설명.
@@ -310,4 +316,7 @@ class CommandFileServiceImpl(
         val extension = originalFileName?.substringAfterLast(".", "") ?: ""
         return if (extension.isNotEmpty()) "$uuid.$extension" else uuid
     }
+
+    fun extractDigits(input: String): Long = input.filter { it.isDigit() }.toLong()
+
 }
